refactor(e2e): rename e2e files for clarity
- Moves contents of `test/*` to `test/e2e/*`
- Renames `test/tsconfig.json` to `test/e2e/tsconfig.wdio.json`
- Removes redundant eslint rules, and fixes eslint errors for `*.e2e.ts` files
  - Updates `tsconfig.eslint.json` to extend `tsconfig.wdio.json`
- Renames e2e test files to end with `.spec.ts` instead of `.e2e.ts`
